+++ Changes.md
@@ -1,7 +1,7 @@
 
 ## 1.0.2 - 2017-09-01
 
-- adds ability to route email via queue.wants and queue.next_hop
+- adds ability to route email via notes.queue.wants and queue.next_hop
 - when destination is a mailbox and next_hop is a LMTP url, routes via LMTP
 
 ## 1.0.1 - 2017-08-29